# Swagger/swagg.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "MovieTicketBooking(MTB) Application",
        "description": "API document for MovieTicketBooking Service",
        "x-task1-appname": "MTB",
        "x-task1-seal": 104437,
        "version": "23.03.0"
    },
...
    "servers": [
        {
            "url": "",
            "description": "http://localhost:1111-Relative server url",
            "variables": {}
        }
    ],
    "paths": {
...
        "/user/{UserID}": {
            "get": {
                "tags": [
                    "USER TABLE"
                ],
                "summary": "Get user by UserID",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "UserID",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the user details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request is not valid",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing Auth Token",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "bbt": [],
                        "token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "USER TABLE"
                ],
                "summary": "Update user details",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "UserID",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User details updated successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request is not valid",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing Auth Token",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "bbt": [],
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "USER TABLE"
                ],
                "summary": "Delete user",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "UserID",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted successfully"
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing Auth Token",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Request is not valid",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "bbt": [],
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "UserID": {
                        "type": "string"
                    },
                    "Name": {
                        "type": "string"
                    },
                    "Password": {
                        "type": "string"
                    },
                    "Email": {
                        "type": "string",
                        "format": "email"
                    },
                    "Phone": {
                        "type": "string"
                    }
                }
            },
...
            "ErrorInfo": {
                "type": "object",
                "properties": {
                    "errorCode": {
                        "type": "string"
                    },
                    "errorMessage": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "bbt": {
                "type": "apiKey",
                "in": "header",
                "name": "BBT"
            },
            "token": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization"
            }
        }
    }
}